k8s-coreos: initial setup for secure etcd cluster
This patch uses the "etcd-ca" tool [1] to generate certificates for the nodes
in k8s cluster.  More specifically, it initializes a local certificate
authority (CA), and create and signs certificates for each node in the cluster.

The certificates will be used for secure peer-to-peer communication in the etcd
cluster, so that the etcd-backed key-value storage can be used reliably. For
this purpose, we first create a node, create certs based on its private IP
address, upload the certs to the VM, add a metadata file so that etcd uses the
certs upon startup, and reboot the machine.

This patch also restricts the etcd client-server communication (e.g., etcdctl)
to the loopback address.

Tested: observed etcd logs to see the cluster is formed correctly; kubelet can
talk to the local etcd; the following command shows a list of minions:
$ curl -X GET http://127.0.0.1:4001/v2/keys/registry/minions

fleet clustering does not work correctly because apparently "etcd redirects
clients from non-leader nodes to the leader ... etcd v0.5 does not behave this
way" [2].

[1] https://github.com/coreos/etcd-ca
[2] https://github.com/coreos/fleet/issues/1015 (see update by bcwaldon@)

Signed-off-by: Amey Deshpande <ameyd@google.com>