k8s-coreos: secure kube-apiserver's serving port
This patch extends the local certificate management to generate certs
for kube-apiserver and its clients (kubectl and kubecfg).  With this patch,
"kube-up.sh" runs to completion, including the validate-cluster step.

Note that this functionality still depends on etcd-ca.  A patch has been added
to etcd-ca so that it can generate client certs.  In a future patch, we will
switch to using easyrsa or openssl.

Signed-off-by: Amey Deshpande <ameyd@google.com>