k8s-coreos: use kube-apiserver's secure port if it is available
In the current version of k8s master on CoreOS, there is no nginx proxying
the well-known https port 443 to kube-apiserver.  Instead, use the port number
that kube-apiserver is actually listening on.  It is assumed that the server is
reachable from outside, i.e., the firewall is configured correctly.

Tested: run the following command with all the required environment variables:

$ ./cluster/kubectl.sh --auth-path=kubernetes_auth get minions

Signed-off-by: Amey Deshpande <ameyd@google.com>