k8s-coreos: fix the bind-addr and published addr of etcd
This patch allows etcd to listen on all interfaces (0.0.0.0) as opposed to
restricting it to the loopback address.  It will publish the private IP of the
VM for cluster formation.

As a result of this patch, we can run etcdctl and fleetctl on the master as
well as the minions without any additional commandline options.

Signed-off-by: Amey Deshpande <ameyd@google.com>